refactor: issue label and make design consistent (#610)
