+++ apps/app/components/core/index.ts
@@ -10,3 +10,4 @@ export * from "./issues-view";
 export * from "./link-modal";
 export * from "./not-authorized-view";
 export * from "./image-picker-popover";
+export * from "./filter-list";
