+++ apps/app/components/core/issues-view.tsx
@@ -9,18 +9,17 @@ import { DragDropContext, DropResult } from "react-beautiful-dnd";
 // services
 import issuesService from "services/issues.service";
 import stateService from "services/state.service";
-import projectService from "services/project.service";
 import modulesService from "services/modules.service";
 // hooks
 import useToast from "hooks/use-toast";
 import useIssuesView from "hooks/use-issues-view";
 // components
-import { AllLists, AllBoards } from "components/core";
+import { AllLists, AllBoards, FilterList } from "components/core";
 import { CreateUpdateIssueModal, DeleteIssueModal } from "components/issues";
 import StrictModeDroppable from "components/dnd/StrictModeDroppable";
 import { CreateUpdateViewModal } from "components/views";
 // ui
-import { Avatar, EmptySpace, EmptySpaceItem, PrimaryButton, Spinner } from "components/ui";
+import { EmptySpace, EmptySpaceItem, PrimaryButton, Spinner } from "components/ui";
 import { CalendarView } from "./calendar-view";
 // icons
 import {
@@ -28,12 +27,10 @@ import {
   PlusIcon,
   RectangleStackIcon,
   TrashIcon,
-  XMarkIcon,
 } from "@heroicons/react/24/outline";
-import { ExclamationIcon, getStateGroupIcon } from "components/icons";
+import { ExclamationIcon } from "components/icons";
 // helpers
 import { getStatesList } from "helpers/state.helper";
-import { replaceUnderscoreIfSnakeCase } from "helpers/string.helper";
 // types
 import {
   CycleIssueResponse,
@@ -49,11 +46,8 @@ import {
   MODULE_ISSUES,
   MODULE_ISSUES_WITH_PARAMS,
   PROJECT_ISSUES_LIST_WITH_PARAMS,
-  PROJECT_ISSUE_LABELS,
-  PROJECT_MEMBERS,
   STATE_LIST,
 } from "constants/fetch-keys";
-import { getPriorityIcon } from "components/icons/priority-icon";
 
 type Props = {
   type?: "issue" | "cycle" | "module";
@@ -111,20 +105,6 @@ export const IssuesView: React.FC<Props> = ({
       : null
   );
   const states = getStatesList(stateGroups ?? {});
-
-  const { data: members } = useSWR(
-    projectId ? PROJECT_MEMBERS(projectId as string) : null,
-    workspaceSlug && projectId
-      ? () => projectService.projectMembers(workspaceSlug as string, projectId as string)
-      : null
-  );
-
-  const { data: issueLabels } = useSWR(
-    projectId ? PROJECT_ISSUE_LABELS(projectId.toString()) : null,
-    workspaceSlug && projectId
-      ? () => issuesService.getIssueLabels(workspaceSlug as string, projectId.toString())
-      : null
-  );
 
   const handleDeleteIssue = useCallback(
     (issue: IIssue) => {
@@ -428,243 +408,7 @@ export const IssuesView: React.FC<Props> = ({
       />
       <div className="mb-5 -mt-4">
         <div className="flex items-center justify-between gap-2">
-          <div className="flex flex-wrap items-center gap-3 text-xs">
-            {Object.keys(filters).map((key) => {
-              if (filters[key as keyof typeof filters] !== null)
-                return (
-                  <div
-                    key={key}
-                    className="flex items-center gap-x-2 rounded-full border bg-white px-2 py-1"
-                  >
-                    <span className="font-medium capitalize text-gray-500">
-                      {replaceUnderscoreIfSnakeCase(key)}:
-                    </span>
-                    {filters[key as keyof IIssueFilterOptions] === null ||
-                    (filters[key as keyof IIssueFilterOptions]?.length ?? 0) <= 0 ? (
-                      <span>None</span>
-                    ) : Array.isArray(filters[key as keyof IIssueFilterOptions]) ? (
-                      <div className="space-x-2">
-                        {key === "state" ? (
-                          <div className="flex items-center gap-x-1">
-                            {filters.state?.map((stateId: any) => {
-                              const state = states?.find((s) => s.id === stateId);
-
-                              return (
-                                <p
-                                  key={state?.id}
-                                  className="inline-flex items-center gap-x-1 rounded-full px-2 py-0.5 font-medium text-white"
-                                  style={{
-                                    color: state?.color,
-                                    backgroundColor: `${state?.color}20`,
-                                  }}
-                                >
-                                  <span>
-                                    {getStateGroupIcon(
-                                      state?.group ?? "backlog",
-                                      "12",
-                                      "12",
-                                      state?.color
-                                    )}
-                                  </span>
-                                  <span>{state?.name ?? ""}</span>
-                                  <span
-                                    className="cursor-pointer"
-                                    onClick={() =>
-                                      setFilters(
-                                        {
-                                          state: filters.state?.filter((s: any) => s !== stateId),
-                                        },
-                                        !Boolean(viewId)
-                                      )
-                                    }
-                                  >
-                                    <XMarkIcon className="h-3 w-3" />
-                                  </span>
-                                </p>
-                              );
-                            })}
-                            <button
-                              type="button"
-                              onClick={() =>
-                                setFilters({
-                                  state: null,
-                                })
-                              }
-                            >
-                              <XMarkIcon className="h-3 w-3" />
-                            </button>
-                          </div>
-                        ) : key === "priority" ? (
-                          <div className="flex items-center gap-x-1">
-                            {filters.priority?.map((priority: any) => (
-                              <p
-                                key={priority}
-                                className={`inline-flex items-center gap-x-1 rounded-full px-2 py-0.5 font-medium capitalize text-white ${
-                                  priority === "urgent"
-                                    ? "bg-red-100 text-red-600 hover:bg-red-100"
-                                    : priority === "high"
-                                    ? "bg-orange-100 text-orange-500 hover:bg-orange-100"
-                                    : priority === "medium"
-                                    ? "bg-yellow-100 text-yellow-500 hover:bg-yellow-100"
-                                    : priority === "low"
-                                    ? "bg-green-100 text-green-500 hover:bg-green-100"
-                                    : "bg-gray-100"
-                                }`}
-                              >
-                                <span>{getPriorityIcon(priority)}</span>
-                                <span>{priority}</span>
-                                <span
-                                  className="cursor-pointer"
-                                  onClick={() =>
-                                    setFilters(
-                                      {
-                                        priority: filters.priority?.filter(
-                                          (p: any) => p !== priority
-                                        ),
-                                      },
-                                      !Boolean(viewId)
-                                    )
-                                  }
-                                >
-                                  <XMarkIcon className="h-3 w-3" />
-                                </span>
-                              </p>
-                            ))}
-                            <button
-                              type="button"
-                              onClick={() =>
-                                setFilters({
-                                  priority: null,
-                                })
-                              }
-                            >
-                              <XMarkIcon className="h-3 w-3" />
-                            </button>
-                          </div>
-                        ) : key === "assignees" ? (
-                          <div className="flex items-center gap-x-1">
-                            {filters.assignees?.map((memberId: string) => {
-                              const member = members?.find((m) => m.member.id === memberId)?.member;
-
-                              return (
-                                <p
-                                  key={memberId}
-                                  className="inline-flex items-center gap-x-1 rounded-full px-0.5 py-0.5 font-medium capitalize"
-                                >
-                                  <Avatar user={member} />
-                                  <span>{member?.first_name}</span>
-                                  <span
-                                    className="cursor-pointer"
-                                    onClick={() =>
-                                      setFilters(
-                                        {
-                                          assignees: filters.assignees?.filter(
-                                            (p: any) => p !== memberId
-                                          ),
-                                        },
-                                        !Boolean(viewId)
-                                      )
-                                    }
-                                  >
-                                    <XMarkIcon className="h-3 w-3" />
-                                  </span>
-                                </p>
-                              );
-                            })}
-                            <button
-                              type="button"
-                              onClick={() =>
-                                setFilters({
-                                  assignees: null,
-                                })
-                              }
-                            >
-                              <XMarkIcon className="h-3 w-3" />
-                            </button>
-                          </div>
-                        ) : (key as keyof IIssueFilterOptions) === "created_by" ? (
-                          <div className="flex items-center gap-x-1">
-                            {filters.created_by?.map((memberId: string) => {
-                              const member = members?.find((m) => m.member.id === memberId)?.member;
-
-                              return (
-                                <p
-                                  key={memberId}
-                                  className="inline-flex items-center gap-x-1 rounded-full px-2 py-0.5 font-medium capitalize"
-                                >
-                                  <Avatar user={member} />
-                                  <span>{member?.first_name}</span>
-                                  <span
-                                    className="cursor-pointer"
-                                    onClick={() =>
-                                      setFilters(
-                                        {
-                                          assignees: filters.created_by?.filter(
-                                            (p: any) => p !== memberId
-                                          ),
-                                        },
-                                        !Boolean(viewId)
-                                      )
-                                    }
-                                  >
-                                    <XMarkIcon className="h-3 w-3" />
-                                  </span>
-                                </p>
-                              );
-                            })}
-                            <button
-                              type="button"
-                              onClick={() =>
-                                setFilters({
-                                  created_by: null,
-                                })
-                              }
-                            >
-                              <XMarkIcon className="h-3 w-3" />
-                            </button>
-                          </div>
-                        ) : key === "labels" ? (
-                          <div className="flex items-center gap-x-1">
-                            {filters.labels?.map((labelId: string) => {
-                              const label = issueLabels?.find((l) => l.id === labelId);
-
-                              if (!label) return null;
-
-                              return (
-                                <div className="flex items-center gap-1">
-                                  <div
-                                    className="h-2 w-2 rounded-full"
-                                    style={{
-                                      backgroundColor:
-                                        label.color && label.color !== "" ? label.color : "#000000",
-                                    }}
-                                  />
-                                  {label.name}
-                                </div>
-                              );
-                            })}
-                            <button
-                              type="button"
-                              onClick={() =>
-                                setFilters({
-                                  labels: null,
-                                })
-                              }
-                            >
-                              <XMarkIcon className="h-3 w-3" />
-                            </button>
-                          </div>
-                        ) : (
-                          (filters[key as keyof IIssueFilterOptions] as any)?.join(", ")
-                        )}
-                      </div>
-                    ) : (
-                      <span className="capitalize">{filters[key as keyof typeof filters]}</span>
-                    )}
-                  </div>
-                );
-            })}
-          </div>
+          <FilterList filters={filters} setFilters={setFilters} />
           {Object.keys(filters).length > 0 &&
             nullFilters.length !== Object.keys(filters).length && (
               <PrimaryButton
@@ -688,24 +432,10 @@ export const IssuesView: React.FC<Props> = ({
               </PrimaryButton>
             )}
         </div>
-        {Object.keys(filters).length > 0 && nullFilters.length !== Object.keys(filters).length && (
-          <button
-            onClick={() =>
-              setFilters({
-                state: null,
-                priority: null,
-                assignees: null,
-                labels: null,
-                created_by: null,
-              })
-            }
-            className="mt-2 flex items-center gap-x-1 rounded-full border bg-white px-3 py-1.5 text-xs"
-          >
-            <span>Clear all filters</span>
-            <XMarkIcon className="h-4 w-4" />
-          </button>
-        )}
       </div>
+
+      <div className="mb-5 border-t" />
+
       <DragDropContext onDragEnd={handleOnDragEnd}>
         <StrictModeDroppable droppableId="trashBox">
           {(provided, snapshot) => (
